Added slightly improved fib.c

+++ others/fib/iter.c
@@ -1,15 +1,15 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-unsigned int fib(unsigned int n)
+long long fib(long long n)
 {
     if (n < 2) return n;
 
-    unsigned int* table = malloc((n + 1) * sizeof(unsigned int));
+    long long* table = malloc((n + 1) * sizeof(long long));
     table[0] = 0;
     table[1] = 1;
-    unsigned int i; for (i = 2; i <= n; i++) table[i] = table[i-2] + table[i - 1];
-    unsigned int res = table[n];
+    long long i; for (i = 2; i <= n; i++) table[i] = table[i-2] + table[i - 1];
+    long long res = table[n];
     free(table);
     return res;
 }
